requirements.txt
annotated-types
anyio
asgiref
attrs
Babel
boto
boto3
botocore
certifi
charset-normalizer
diff-match-patch
distro
dj-database-url
Django
django-colorfield
django-cors-headers
django-filter
django-import-export
django-phonenumber-field
django-rest-auth
django-storages
django-tinymce
djangorestframework
drf-spectacular
exceptiongroup
filelock
gunicorn
h11
httpcore
httpx
idna
inflection
jmespath
jsonschema
jsonschema-specifications
openai
packaging
phonenumberslite
pillow
psycopg
psycopg2-binary
pydantic
pydantic_core
python-dateutil
python-decouple
PyYAML
referencing
requests
requests-file
rpds-py
s3transfer
six
sniffio
sqlparse
tablib
tldextract
tqdm
typing_extensions
uritemplate
urllib3
whitenoise
